% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt.R
\name{irt}
\alias{irt}
\title{Bayesian Item Response Theory Model}
\usage{
irt(formula, chi = 0, data, returns = "samples")
}
\arguments{
\item{formula}{A formula like score ~ person + item. The lhs is the binary
score or accuracy variable. The rhs gives the person
(student/respondent/subject/participant) and item (question) variables, in
that order.}

\item{chi}{Guessing parameter: 0 < c < 1.}

\item{data}{The data frame with the required variables}

\item{returns}{Which object or objects is returned. Options are 'samples',
'model', or 'both'. The 'samples' is a list of 'person' and 'item' related
parameter samples. The 'model' is the rstan model. The 'both' options
return a list with both 'samples' and 'model.}
}
\value{
An rstan model
}
\description{
Implements a Bayesian multilevel 2PL IRT model
}
\examples{
\dontrun{
M <- irt(score ~ nid + item, data = exam_df)
}
}
