% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtb-package.R
\docType{package}
\name{irtb-package}
\alias{irtb}
\alias{irtb-package}
\title{irtb: Bayesian Item Response Theory Models Using Stan}
\description{
High level functions for doing Bayesian Item Response Theory (IRT) analysis with Stan.
}
\author{
\strong{Maintainer}: Mark Andrews \email{mark.andrews@ntu.ac.uk}

}
\keyword{internal}
